refactor(web): extract ChannelDeleteFlow component
